Rollback change to reuse_manager.py
Cahnge to reuse_manager.py is postponed till after this PR.

Review comment:

https://github.com/vllm-project/vllm/pull/40538#discussion_r3129980578
